Fixed typo & directory refactoring (#264)
* Fixed typo & directory refactoring

* Typo 'questionnaire'

* Class name changes according to typo
