+++ lib/routes.dart
@@ -16,9 +16,9 @@
 */
 
 import 'package:flutter/cupertino.dart';
-import 'package:picos/screens/fogot_password_screen/forgot_password_screen.dart';
-import 'package:picos/screens/follow_up_screen/edit_follow_up_screen.dart';
-import 'package:picos/screens/follow_up_screen/follow_up_list.dart';
+import 'package:picos/screens/forgot_password_screen/forgot_password_screen.dart';
+import 'package:picos/screens/study_nurse_screen/follow_up_screen/edit_follow_up_screen.dart';
+import 'package:picos/screens/study_nurse_screen/follow_up_screen/follow_up_list.dart';
 import 'package:picos/screens/legals_screen.dart';
 import 'package:picos/screens/login_screen.dart';
 import 'package:picos/screens/my_documents_screen/add_document_screen.dart';
@@ -32,7 +32,7 @@ import 'package:picos/screens/family_member_list_screen/add_family_member_screen
 import 'package:picos/screens/family_member_list_screen/family_members_screen.dart';
 import 'package:picos/screens/privacy_notice_screen.dart';
 import 'package:picos/screens/profile_screen/profile_screen.dart';
-import 'package:picos/screens/questionaire_screen/questionaire_screen.dart';
+import 'package:picos/screens/questionnaire_screen/questionnaire_screen.dart';
 import 'package:picos/screens/study_nurse_screen/menu_screen/edit_patient_screen.dart';
 import 'package:picos/screens/study_nurse_screen/menu_screen/menu_main_screen.dart';
 import 'package:picos/screens/study_nurse_screen/catalog_of_items_screen/catalog_of_items_screen.dart';
@@ -85,7 +85,7 @@ class Routes {
           const MyVisitsScreen(),
       '/login-screen': (BuildContext ctx) => const LoginScreen(),
       '/questionnaire-screen/questionnaire-screen': (BuildContext ctx) =>
-          const QuestionaireScreen(),
+          const QuestionnaireScreen(),
       '/my-documents-screen/my-documents': (BuildContext ctx) =>
           const MyDocumentsScreen(),
       '/my-documents-screen/add-documents': (BuildContext ctx) =>
