+++ lib/screens/questionnaire_screen/widgets/sleep_quality_card.dart
@@ -16,7 +16,7 @@
 */
 
 import 'package:flutter/material.dart';
-import 'package:picos/screens/questionaire_screen/widgets/questionaire_card.dart';
+import 'package:picos/screens/questionnaire_screen/widgets/questionnaire_card.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../../../widgets/picos_label.dart';
@@ -253,7 +253,7 @@ class _SleepQualityCardState extends State<SleepQualityCard> {
       },
     );
 
-    return QuestionaireCard(
+    return QuestionnaireCard(
       padding: const EdgeInsets.symmetric(
         vertical: 15,
         horizontal: 0,
