+++ lib/screens/questionnaire_screen/widgets/text_field_card.dart
@@ -17,7 +17,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:picos/screens/questionaire_screen/widgets/questionaire_card.dart';
+import 'package:picos/screens/questionnaire_screen/widgets/questionnaire_card.dart';
 import 'package:picos/util/replace_comma_formatter.dart';
 
 import '../../../widgets/picos_label.dart';
@@ -62,7 +62,7 @@ class TextFieldCard extends StatelessWidget {
       initialStringValue = initialValue.toString();
     }
 
-    return QuestionaireCard(
+    return QuestionnaireCard(
       label: PicosLabel(label),
       child: PicosTextField(
         initialValue: initialStringValue,
